pyproject.toml
[project]
name = "simp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-template-partials>=24.4",
    "django-widget-tweaks>=1.5.0",
    "django>=5.2.4",
    "granian[reload,uvloop]>=2.5.0",
    "markitdown[all]>=0.1.2",
    "meilisearch>=0.36.0",
    "ollama>=0.5.1",
    "psycopg[binary,pool]>=3.2.9",
    "python-socketio[asyncio-client]>=5.13.0",
    "pyzmq>=27.0.0",
]

[tool.uv]
dev-dependencies = [
    "daphne>=4.2.1",
    "ipython>=9.4.0",
    "jupyterlab>=4.4.5",
    "rich>=14.1.0",
    "ruff>=0.12.5",
    "urllib3>=2.5.0",
    "watchfiles>=1.1.0",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py39" # Or your project's Python version

[tool.ruff.lint]
select = ["E", "F"] # Pyflakes (F) and a subset of pycodestyle (E)
